requirements.txt
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
babel==2.17.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
Django==5.1.6
django-debug-toolbar==5.0.1
docutils==0.21.2
factory_boy==3.3.3
Faker @ file:///home/conda/feedstock_root/build_artifacts/faker_1739265284709/work
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0rc1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imagesize==1.4.1
iniconfig @ file:///home/linux1/recipes/ci/iniconfig_1610983019677/work
Jinja2==3.1.5
jiter==0.8.2
# Editable Git install with no remote (backend==0.1.0)
-e /Users/jacobhornsvennevik/Documents/GitHub/backend
MarkupSafe==3.0.2
openai==1.62.0
packaging @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_15t4xe1fp0/croot/packaging_1734472125760/work
pillow==11.1.0
pluggy @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_70ykkrsb12/croot/pluggy_1733169619735/work
proto-plus==1.26.0
protobuf==5.29.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyMuPDF==1.25.3
pymupdf4llm==0.0.17
pyparsing==3.2.1
pytest==8.3.4
pytest-django==4.10.0
pytest-mock==3.14.0
requests==2.32.3
rsa==4.9
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==8.1.3
sphinx-autobuild==2024.10.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.3
starlette==0.45.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1737541069190/work
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==14.2
